feat([RA-PRO.14]): columna canónica analytics.sales_daily.units_base (venta en CAJAS)
Materializa la unidad canónica en vez de inferirla en cada lectura. Regla por ORIGEN
(confirmada con datos): Wincaja (wincaja_*) vende CAJAS → units tal cual; Kepler POS
(tienda/credito) vende PIEZAS sueltas y a veces cajas → resuelve por fila (pieza si
revenue/units < cost_with_tax/√factor_sale → units/factor_sale).

- mig 20260724120000: analytics.sales_daily.units_base numeric (nullable, idempotente).
- import-sales-units-base.js: normalizador idempotente (ventana 400d). Dry-run vs prod:
  tienda 770M pz → 2.98M cajas (colapso 767M), Wincaja colapso 0 (ya cajas).
- import-inventory-health: lee COALESCE(units_base, <regla inline fallback>) → demanda en cajas.
- run-prod-feeds: normalizador corre tras sales-fact y antes de inventory-health.

PENDIENTE PROD: aplicar mig 20260724120000 + correr import-sales-units-base --apply +
import-inventory-health --apply + import-computed-reorder --apply + redeploy api+view.

+++ database/importers/kepler/import-sales-units-base.js
@@ -51,7 +51,7 @@ const RULE = `
              round(sum(sd.units) - sum(${RULE})) AS colapso
         FROM analytics.sales_daily sd
         JOIN catalog.products p ON p.tenant_id = sd.tenant_id AND p.id = sd.product_id
-       WHERE sd.tenant_id = $1 AND sd.sale_date >= current_date - $2
+       WHERE sd.tenant_id = $1 AND sd.sale_date >= current_date - $2::int
        GROUP BY sd.channel ORDER BY colapso DESC NULLS LAST`, [M, LOOKBACK]);
     console.table(imp);
 
@@ -64,7 +64,7 @@ const RULE = `
          SET units_base = ${RULE}
         FROM catalog.products p
        WHERE sd.tenant_id = $1 AND p.tenant_id = sd.tenant_id AND p.id = sd.product_id
-         AND sd.sale_date >= current_date - $2
+         AND sd.sale_date >= current_date - $2::int
          AND sd.units_base IS DISTINCT FROM (${RULE})`, [M, LOOKBACK]);
     await db.query('COMMIT');
     console.log(`\n[APPLY] COMMIT — ${res.rowCount} filas normalizadas.`);
